+++ src/idvc/dvc_interface.py
@@ -20,13 +20,11 @@ import math
 
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
-from matplotlib.figure import Figure
 import matplotlib.pyplot as plt
 
 from functools import partial
 from datetime import datetime
 
-from os import listdir
 
 import vtk
 from ccpi.viewer import viewer2D, viewer3D
@@ -59,16 +57,10 @@ import tempfile
 import json
 import shutil
 import zipfile
-import zlib
 
-import csv
 from functools import reduce
 
-import subprocess
-
 import copy
-
-from distutils.dir_util import copy_tree
 
 from idvc.io import ImageDataCreator
 
@@ -455,7 +447,7 @@ class MainWindow(QMainWindow):
             "Once you are satisfied with the registration, make sure the point 0 you have selected is the point you want the DVC to start from."
             )
         
-        self.help_text.append("Enable trace mode by clicking on the 2D viewer, then pressing 't'. Then you may draw a region freehand.\n"
+        self.help_text.append("To create a mask you need to create a selection. Start tracing a freehand region for the selection by clicking 'Start Tracing' button.\n"
             "When you are happy with your region click 'Create Mask'.")
 
         self.help_text.append("Dense point clouds that accurately reflect sample geometry and reflect measurement objectives yield the best results.\n"
@@ -614,7 +606,9 @@ class MainWindow(QMainWindow):
                         return #prevents dialog showing for every single file by exiting the for loop
                 image[image_var] = filenames
                 if label is not None:
-                    label.setText(os.path.basename(self.image[image_var][0]) + " + " + str(len(files)) + " more files.")
+                    label.setText(
+                        os.path.dirname(self.image[image_var][0]) + "\n" +\
+                        os.path.basename(self.image[image_var][0]) + " + " + str(len(files)) + " more files.")
 
             if next_button is not None:
                 next_button.setEnabled(True)
@@ -657,7 +651,10 @@ class MainWindow(QMainWindow):
         msg.setIcon(QMessageBox.Critical)
         msg.setWindowTitle(title)
         msg.setText(message)
-        msg.setDetailedText(self.e.ErrorMessage())
+        try:
+            msg.setDetailedText(self.e.ErrorMessage())
+        except:
+            pass
         if action_button is not None:
             msg.addButton(action_button, msg.ActionRole)
         msg.exec_()
@@ -813,6 +810,7 @@ class MainWindow(QMainWindow):
                 self.PointCloudWorker("load pointcloud file")
                 self.pointCloudLoaded = True
                 self.no_mask_pc_load = False
+                self.pointcloud_is = 'loaded'
 
         if(self.reg_load):
             self.displayRegistrationViewer(registration_open = True)
@@ -1586,7 +1584,7 @@ It is used as a global starting point and a translation reference."
         else:
             if not (hasattr(self, 'unsampled_ref_image_data') and hasattr(self, 'unsampled_corr_image_data')):
                 self.unsampled_ref_image_data = self.ref_image_data 
-                self.LoadCorrImageForReg()
+                self.LoadCorrImageForReg(crop_corr_image=True)
             else:
                 self.completeRegistration()
 
@@ -1867,15 +1865,26 @@ It is used as a global starting point and a translation reference."
         mp_widgets['extendMaskCheck'].setText("Extend mask")
         mp_widgets['extendMaskCheck'].setToolTip("You may draw a second trace. Select extend mask to extend the mask to this second traced region.")
         mp_widgets['extendMaskCheck'].setEnabled(False)
+        mp_widgets['extendMaskCheck'].stateChanged.connect(self.warnIfUnchecking)
 
         formLayout.setWidget(widgetno,QFormLayout.FieldRole, mp_widgets['extendMaskCheck'])
+        widgetno += 1
+
+        # Add start tracing button
+        mp_widgets['start_tracing'] = QPushButton(groupBox)
+        mp_widgets['start_tracing'].setText("Start Tracing")
+        mp_widgets['start_tracing'].clicked.connect(self.ToggleTracing)
+        mp_widgets['start_tracing'].setCheckable(True)
+        mp_widgets['start_tracing'].setChecked(False)
+        mp_widgets['start_tracing'].setToolTip("Begin drawing a region on the viewer to create a mask.")
+        formLayout.setWidget(widgetno, QFormLayout.FieldRole, mp_widgets['start_tracing'])
         widgetno += 1
 
         # Add submit button
         mp_widgets['submitButton'] = QPushButton(groupBox)
         mp_widgets['submitButton'].setText("Create Mask")
         mp_widgets['submitButton'].clicked.connect(lambda: self.MaskWorker("extend"))
-        mp_widgets['submitButton'].setToolTip("Press 't' and draw a region on the viewer to create a mask.")
+        mp_widgets['submitButton'].setToolTip("Press 'Start Tracing' button and draw a region on the viewer to create a mask.")
         formLayout.setWidget(widgetno, QFormLayout.FieldRole, mp_widgets['submitButton'])
         widgetno += 1
 
@@ -1897,6 +1906,31 @@ It is used as a global starting point and a translation reference."
 
         # Add elements to layout
         self.addDockWidget(QtCore.Qt.LeftDockWidgetArea, dockWidget)
+
+
+    def warnIfUnchecking(self):
+        if not self.mask_parameters['extendMaskCheck'].isChecked() and self.mask_parameters['extendMaskCheck'].isEnabled():
+            self.warningDialog(window_title="Attention", 
+                               message="If you do not clear the mask and draw again the app will be very sad!" )
+
+
+    def ToggleTracing(self):
+        '''Toggles the tracing widget for tracing the mask'''
+        # notice that the viewer will capture the keypress event linked with activating the tracing
+        viewer = self.vis_widget_2D.getViewer()
+        mp_widgets = self.mask_parameters
+        # we come here after the user has clicked on the button:
+        # thins means that if the button is checked, it was not before,
+        # So the user expects to start tracing
+        if mp_widgets['start_tracing'].isChecked():
+            # enable tracing
+            mp_widgets['start_tracing'].setText("Stop Tracing")
+            viewer.imageTracer.On()
+        else:
+            # disable tracing
+            mp_widgets['start_tracing'].setText("Start Tracing")
+            viewer.imageTracer.Off()
+
 
     def MaskWorker(self, type):
         v = self.vis_widget_2D.frame.viewer
@@ -1922,7 +1956,11 @@ It is used as a global starting point and a translation reference."
             
         self.create_progress_window("Loading", "Loading Mask")
         self.mask_worker.signals.progress.connect(self.progress)
-       
+
+        # disable tracing on the viewer and
+        # set the button to unchecked and text to start tracing
+        self._disableTracingAndResetUI()
+
         self.progress_window.setValue(10)
         self.threadpool.start(self.mask_worker)  
         self.mask_worker.signals.error.connect(self.select_mask)
@@ -1940,8 +1978,9 @@ It is used as a global starting point and a translation reference."
         #we can easily get the image data v.image2 but would need a stencil?
 
         # print("Extend mask")
-        progress_callback = kwargs.get('progress_callback', None)
-        
+        progress_callback = kwargs.get('progress_callback', lambda x: logging.info("extendMask Progress: {}".format(x)))
+        # setup the progress bar
+        ui_update = partial(self.ui_progress_update, progress_callback)
         v = self.vis_widget_2D.frame.viewer
 
         poly = vtk.vtkPolyData()
@@ -1975,7 +2014,7 @@ It is used as a global starting point and a translation reference."
 
         #print(image_data.GetSpacing())
         
-        progress_callback.emit(40)
+        # progress_callback.emit(40)
 
 
         mask0 = Converter.numpy2vtkImage(np.zeros((dims[0],dims[1],dims[2]),order='F', 
@@ -1993,10 +2032,12 @@ It is used as a global starting point and a translation reference."
         stencil.SetBackgroundInputData(mask0)
 
         stencil.SetStencilConnection(lasso.GetOutputPort())
+
+        stencil.AddObserver(vtk.vtkCommand.ProgressEvent, ui_update)
         stencil.Update()
         dims = stencil.GetOutput().GetDimensions()
 
-        progress_callback.emit(80)
+        # progress_callback.emit(80)
 
         #print("Stencil dims: " + str(dims))
 
@@ -2010,7 +2051,7 @@ It is used as a global starting point and a translation reference."
         #vtkutils.copyslices(stencil.GetOutput(), sliceno , zmin, zmax, orientation, None)
         stencil_output = self.copySlices(stencil.GetOutput(), sliceno , zmin, zmax, orientation, None)
 
-        progress_callback.emit(85)
+        # progress_callback.emit(85)
 
         # save the mask to a file in temp folder
         writer = vtk.vtkMetaImageWriter()
@@ -2018,11 +2059,11 @@ It is used as a global starting point and a translation reference."
         writer.SetFileName(os.path.join(tmpdir, "Masks", "latest_selection.mha"))
         self.mask_file = "Masks/latest_selection.mha"
 
-        progress_callback.emit(90)
+        # progress_callback.emit(90)
 
         # if extend mask -> load temp saved mask
         if self.mask_parameters['extendMaskCheck'].isChecked():
-            self.setStatusTip('Extending mask')
+            # self.setStatusTip('Extending mask')
             if os.path.exists(os.path.join(tmpdir, "Masks", "latest_selection.mha")):
                 # print  ("extending mask ", os.path.join(tmpdir, "Masks/latest_selection.mha"))
                 reader = vtk.vtkMetaImageReader()
@@ -2036,11 +2077,13 @@ It is used as a global starting point and a translation reference."
                 math.SetInput2Data(reader.GetOutput())
                 math.Update()
 
+                progress_callback.emit(1)
                 threshold = vtk.vtkImageThreshold()
                 threshold.ThresholdBetween(1, 255)
                 threshold.ReplaceInOn()
                 threshold.SetInValue(1)
                 threshold.SetInputConnection(math.GetOutputPort())
+                threshold.AddObserver(vtk.vtkCommand.ProgressEvent, ui_update)
                 threshold.Update()
 
                 writer.SetInputData(threshold.GetOutput())
@@ -2051,11 +2094,28 @@ It is used as a global starting point and a translation reference."
             writer.SetInputData(stencil_output)
             self.mask_data = stencil_output
 
+        # the writer only updates two values 0 and 1
+        # also it blocks the interface (why???)
+        # adding a pause allows the progress to update correctly
+        writer.AddObserver(vtk.vtkCommand.ProgressEvent, ui_update)
+
+        progress_callback.emit(99)
+        import time
+        time.sleep(1)
         writer.Write() # writes to file.
         self.mask_parameters['extendMaskCheck'].setEnabled(True)
-        self.setStatusTip('Done')
+        # self.setStatusTip('Done')
 
-        progress_callback.emit(100)
+        # progress_callback.emit(100)
+        self._disableTracingAndResetUI()
+
+    def _disableTracingAndResetUI(self):
+        # disable tracing on the viewer
+        v = self.vis_widget_2D.frame.viewer
+        v.imageTracer.Off()
+        # set the button to unchecked and text to start tracing
+        self.mask_parameters['start_tracing'].setChecked(False)
+        self.mask_parameters['start_tracing'].setText("Start Tracing")
 
     def copySlices(self, indata, fromslice, min, max, orientation, progress_callback):
 
@@ -2084,9 +2144,8 @@ It is used as a global starting point and a translation reference."
         load_session = kwargs.get('load_session', False)
         progress_callback = kwargs.get('progress_callback', None)
         time.sleep(0.1) #required so that progress window displays
-        progress_callback.emit(30)
-        #Appropriate modification to Point Cloud Panel
-        self.updatePointCloudPanel()
+        # progress_callback.emit(30)
+        
         
         v =  self.vis_widget_2D.frame.viewer
         if not isinstance(v.img3D, vtk.vtkImageData):
@@ -2099,7 +2158,7 @@ It is used as a global starting point and a translation reference."
         tmpdir = tempfile.gettempdir()
         if (load_session):
             self.mask_reader.SetFileName(os.path.join(tmpdir, "Masks", "latest_selection.mha"))
-            progress_callback.emit(40)
+            # progress_callback.emit(40)
         else:
             filename = self.mask_parameters["masksList"].currentText()
             self.mask_reader.SetFileName(os.path.join(tmpdir, "Masks", filename))
@@ -2116,16 +2175,20 @@ It is used as a global starting point and a translation reference."
                 self.axis = axis
 
                 self.sliceno = self.mask_details[filename][1]
-            progress_callback.emit(60)
-            
+            # progress_callback.emit(60)
+        
+        ui_update = partial(self.ui_progress_update, progress_callback)
+        self.mask_reader.AddObserver(vtk.vtkCommand.ProgressEvent, ui_update)
+        
         self.mask_reader.Update()
-        #progress_callback.emit(50)
+        progress_callback.emit(50)
 
         writer = vtk.vtkMetaImageWriter()
        
         writer.SetFileName(os.path.join(tmpdir, "Masks", "latest_selection.mha"))
         writer.SetInputConnection(self.mask_reader.GetOutputPort())
-        progress_callback.emit(80)
+        # progress_callback.emit(80)
+        writer.AddObserver(vtk.vtkCommand.ProgressEvent, ui_update)
         writer.Write()
         self.mask_file = "Masks/latest_selection.mha"
         
@@ -2136,8 +2199,16 @@ It is used as a global starting point and a translation reference."
         if not dims == self.mask_reader.GetOutput().GetDimensions():
             #print("Not compatible")
             return 
-
         self.mask_data = self.mask_reader.GetOutput()
+        progress_callback.emit(100)
+
+    def ui_progress_update(self, callback, vtkcaller, event):
+        '''connects the progress from a vtkAlgorithm to a Qt callback from eqt Worker'''
+        print ("{} progress {}".format(type(vtkcaller), vtkcaller.GetProgress()))
+        progress = vtkcaller.GetProgress()*100-1
+        if progress < 0:
+            progress = 0
+        callback.emit(progress)
 
     def select_mask(self): 
         dialogue = QFileDialog()
@@ -2156,6 +2227,7 @@ It is used as a global starting point and a translation reference."
 
     def clearMask(self):
         self.mask_parameters['extendMaskCheck'].setEnabled(False)
+        self.mask_parameters['extendMaskCheck'].setChecked(False)
         self.mask_parameters['submitButton'].setText("Create Mask")
         self.vis_widget_2D.frame.viewer.setInputData2(vtk.vtkImageData()) #deletes mask
         self.mask_reader = None
@@ -2167,12 +2239,21 @@ It is used as a global starting point and a translation reference."
         #self.lasso = vtk.vtkLassoStencilSource()
         #self.vis_widget_2D.frame.viewer.imageTracer = vtk.vtkImageTracerWidget() #cause problems - as removes functionality
         #self.vis_widget_2D.frame.viewer.imageTracer.Modified()
-
+    
     def DisplayMask(self, type = None):
+        #Appropriate modification to Point Cloud Panel
+        self.updatePointCloudPanel()
+
         self.mask_parameters['extendMaskCheck'].setEnabled(True)
         v = self.vis_widget_2D.frame.viewer
         if (hasattr(self,'mask_data')):
+            ui_update = partial(lambda dialog, vtkcaller, event: dialog.setValue(vtkcaller.GetProgress()*100), self.progress_window)
+            v.imageSlice.AddObserver(vtk.vtkCommand.ProgressEvent, ui_update)
             v.setInputData2(self.mask_data)
+            # we loaded the mask, the app crashes if someone clicks on create mask
+            # without clearing the mask first unless extendMaskCheck is checked 
+            # Forcibly check extend mask checkbox
+            self.mask_parameters['extendMaskCheck'].setChecked(True)
         else:
             if v.img3D:
                 self.warningDialog( 
@@ -2285,11 +2366,18 @@ It is used as a global starting point and a translation reference."
         pc['pointcloud_volume_shape_entry'] = self.subvolumeShapeValue
 
         # Add horizonal seperator
+        # Generate panel
         self.seperator = QFrame(self.graphParamsGroupBox)
         self.seperator.setFrameShape(QFrame.HLine)
         self.seperator.setFrameShadow(QFrame.Raised)
         self.graphWidgetFL.setWidget(widgetno, QFormLayout.SpanningRole, self.seperator)
         widgetno += 1
+        # Load point cloud section 
+        # add a separator and title
+        generatePointCloudLabel = QLabel("Generate Pointcloud", self.graphParamsGroupBox)
+        self.graphWidgetFL.setWidget(widgetno, QFormLayout.LabelRole, generatePointCloudLabel)
+        widgetno += 1
+
 
         # Add collapse priority field
         self.dimensionalityLabel = QLabel(self.graphParamsGroupBox)
@@ -2332,7 +2420,7 @@ A 3D pointcloud is created within the full extent of the mask.")
         self.overlapXValueEntry.setMaximum(0.99)
         self.overlapXValueEntry.setMinimum(0.00)
         self.overlapXValueEntry.setSingleStep(0.01)
-        self.overlapXValueEntry.setSizePolicy(QSizePolicy.MinimumExpanding, QSizePolicy.MinimumExpanding)
+        self.overlapXValueEntry.setSizePolicy(QSizePolicy.MinimumExpanding, QSizePolicy.Minimum)
         self.overlapXValueEntry.valueChanged.connect(self.displaySubvolumePreview)
         self.overlapXValueEntry.setToolTip(overlap_tooltip_text)
         if orientation == 0:
@@ -2351,7 +2439,7 @@ A 3D pointcloud is created within the full extent of the mask.")
         self.overlapYValueEntry.setMaximum(0.99)
         self.overlapYValueEntry.setMinimum(0.00)
         self.overlapYValueEntry.setSingleStep(0.01)
-        self.overlapYValueEntry.setSizePolicy(QSizePolicy.MinimumExpanding, QSizePolicy.MinimumExpanding)
+        self.overlapYValueEntry.setSizePolicy(QSizePolicy.MinimumExpanding, QSizePolicy.Minimum)
         self.overlapYValueEntry.valueChanged.connect(self.displaySubvolumePreview)
         self.overlapYValueEntry.setToolTip(overlap_tooltip_text)
         if orientation == 1:
@@ -2370,7 +2458,7 @@ A 3D pointcloud is created within the full extent of the mask.")
         self.overlapZValueEntry.setMaximum(0.99)
         self.overlapZValueEntry.setMinimum(0.00)
         self.overlapZValueEntry.setSingleStep(0.01)
-        self.overlapZValueEntry.setSizePolicy(QSizePolicy.MinimumExpanding, QSizePolicy.MinimumExpanding)
+        self.overlapZValueEntry.setSizePolicy(QSizePolicy.MinimumExpanding, QSizePolicy.Minimum)
         self.overlapZValueEntry.valueChanged.connect(self.displaySubvolumePreview)
         self.overlapZValueEntry.setToolTip(overlap_tooltip_text)
         if orientation == 2:
@@ -2477,7 +2565,7 @@ A 3D pointcloud is created within the full extent of the mask.")
         # Add submit button
         self.graphParamsSubmitButton = QPushButton(self.graphParamsGroupBox)
         self.graphParamsSubmitButton.setText("Generate Point Cloud")
-        self.graphParamsSubmitButton.clicked.connect(lambda: self.createSavePointCloudWindow(save_only=False))
+        self.graphParamsSubmitButton.clicked.connect(self._generatePointCloudClicked)
         self.graphWidgetFL.setWidget(widgetno, QFormLayout.FieldRole, self.graphParamsSubmitButton)
         widgetno += 1
         # Add elements to layout
@@ -2485,6 +2573,17 @@ A 3D pointcloud is created within the full extent of the mask.")
         self.graphDockVL.addWidget(self.dockWidget)
         self.pointCloudDockWidget.setWidget(self.pointCloudDockWidgetContents)
         self.addDockWidget(QtCore.Qt.LeftDockWidgetArea, self.pointCloudDockWidget)
+        widgetno += 1
+
+        # Load point cloud section 
+        # add a separator and title
+        seperator = QFrame(self.graphParamsGroupBox)
+        seperator.setFrameShape(QFrame.HLine)
+        seperator.setFrameShadow(QFrame.Raised)
+        self.graphWidgetFL.setWidget(widgetno, QFormLayout.SpanningRole, seperator)
+        widgetno += 1
+        generatePointCloudLabel = QLabel("Load Pointcloud", self.graphParamsGroupBox)
+        self.graphWidgetFL.setWidget(widgetno, QFormLayout.LabelRole, generatePointCloudLabel)
         widgetno += 1
 
         pc['pointcloudList'] = QComboBox(self.graphParamsGroupBox)
@@ -2515,6 +2614,17 @@ The first point is significant, as it is used as a global starting point and ref
         self.graphWidgetFL.setWidget(widgetno, QFormLayout.FieldRole, pc['roi_browse'])
         widgetno += 1
 
+        # Display point cloud section 
+        # add a separator and title
+        seperator = QFrame(self.graphParamsGroupBox)
+        seperator.setFrameShape(QFrame.HLine)
+        seperator.setFrameShadow(QFrame.Raised)
+        self.graphWidgetFL.setWidget(widgetno, QFormLayout.SpanningRole, seperator)
+        widgetno += 1
+        generatePointCloudLabel = QLabel("Display Pointcloud", self.graphParamsGroupBox)
+        self.graphWidgetFL.setWidget(widgetno, QFormLayout.LabelRole, generatePointCloudLabel)
+        widgetno += 1
+
         pc['clear_button'] = QPushButton(self.graphParamsGroupBox)
         pc['clear_button'].setText("Clear Point Cloud")
         pc['clear_button'].clicked.connect(self.clearPointCloud)
@@ -2523,7 +2633,7 @@ The first point is significant, as it is used as a global starting point and ref
 
         pc['subvolumes_check'] = QCheckBox(self.graphParamsGroupBox)
         pc['subvolumes_check'].setText("Display Subvolume Regions")
-        pc['subvolumes_check'].setChecked(True)
+        pc['subvolumes_check'].setChecked(False)
         pc['subvolumes_check'].stateChanged.connect( partial(self.showHideActor,actor_name='subvol_actor') )
         self.graphWidgetFL.setWidget(widgetno, QFormLayout.FieldRole, pc['subvolumes_check'])
         widgetno += 1
@@ -2541,6 +2651,10 @@ The first point is significant, as it is used as a global starting point and ref
         pc['pc_points_value'] = QLabel("0")
 
         self.graphWidgetFL.setWidget(widgetno, QFormLayout.FieldRole, pc['pc_points_value'])
+
+    def _generatePointCloudClicked(self):
+        self.pointcloud_is = 'generated'
+        self.createSavePointCloudWindow(save_only=False)
 
     def displaySubvolumePreview(self):
         if self.pointcloud_parameters['subvolume_preview_check'].isChecked():
@@ -2673,8 +2787,8 @@ The first point is significant, as it is used as a global starting point and ref
 
     def select_pointcloud(self): #, label):
         dialogue = QFileDialog()
+        self.roi = None
         self.roi = dialogue.getOpenFileName(self,"Select a roi")[0]
-        #print(self.roi)
         if self.roi:
             self.PointCloudWorker("load pointcloud file")
 
@@ -2684,6 +2798,9 @@ The first point is significant, as it is used as a global starting point and ref
             self.roi = os.path.abspath(os.path.join(tempfile.tempdir, filename))
             self.pointcloud_parameters['pointcloudList'].addItem(filename)
             self.pointcloud_parameters['pointcloudList'].setCurrentText(filename)
+        
+        if self.roi is not None:
+            self.pointcloud_is = 'loaded'
 
 
     def PointCloudWorker(self, type, filename = None, disp_file = None, vector_dim = None):
@@ -3079,13 +3196,17 @@ Please select a replacement pointcloud file.')
     def DisplayNumberOfPointcloudPoints(self):
         # print("Update DisplayNumberOfPointcloudPoints to ", self.pc_no_points)
         self.pointcloud_parameters['pc_points_value'].setText(str(self.pc_no_points))
-        self.result_widgets['pc_points_value'].setText(str(self.pc_no_points))
         self.rdvc_widgets['run_points_spinbox'].setMaximum(int(self.pc_no_points))
+        if hasattr(self, 'num_processed_points'):
+            self.result_widgets['pc_points_value'].setText(str(self.num_processed_points))
         
 
     def DisplayLoadedPointCloud(self):
         self.setup2DPointCloudPipeline()
         self.setup3DPointCloudPipeline()
+        # hide actor if user does not request to see it. Off by default
+        self.showHideActor(self.pointcloud_parameters['subvolumes_check'].isChecked(), actor_name='subvol_actor')
+        
         #Update window so pointcloud is instantly visible without user having to interact with viewer first
         self.vis_widget_2D.frame.viewer.getRenderWindow().Render()
         self.vis_widget_3D.frame.viewer.getRenderWindow().Render()
@@ -3097,6 +3218,7 @@ Please select a replacement pointcloud file.')
         self.pointCloudLoaded = True
         self.pointCloud_details["latest_pointcloud.roi"] = [self.pointCloud_subvol_size, self.pointCloud_overlap, self.pointCloud_rotation, self.pointCloud_shape]
         self.DisplayNumberOfPointcloudPoints()
+        self.pointcloud_is = 'loaded'
         
 
     def DisplayPointCloud(self):
@@ -3162,6 +3284,9 @@ Try modifying the subvolume size before creating a new pointcloud, and make sure
 
         self.progress_window.setValue(100)
 
+        # hide actor if user does not request to see it. Off by default
+        self.showHideActor(self.pointcloud_parameters['subvolumes_check'].isChecked(), actor_name='subvol_actor')
+        
         self.warningDialog(window_title="Success", message="Point cloud created." )
         self.pointCloud_details["latest_pointcloud.roi"] = [self.pointCloud_subvol_size, self.pointCloud_overlap, self.pointCloud_rotation, self.pointCloud_shape]
         self.DisplayNumberOfPointcloudPoints()
@@ -3229,7 +3354,7 @@ Try modifying the subvolume size before creating a new pointcloud, and make sure
         displ = self.loadDisplacementFile(disp_file, disp_wrt_point0 = self.result_widgets['vec_entry'].currentIndex() == 2, \
                                                      multiplier = self.result_widgets['scale_vectors_entry'].value())
 
-        self.pc_no_points = np.shape(displ)[0]
+        self.num_processed_points = np.shape(displ)[0]
         self.DisplayNumberOfPointcloudPoints()
 
         logging.info('Adding vectors 2D')
@@ -3797,7 +3922,7 @@ This parameter has a strong effect on computation time, so be careful."
         rdvc_widgets['subvol_points_spinbox'] = QSpinBox(singleRun_groupBox)
         rdvc_widgets['subvol_points_spinbox'].setMinimum(100)
         rdvc_widgets['subvol_points_spinbox'].setMaximum(50000)
-        rdvc_widgets['subvol_points_spinbox'].setMaximum(10000)
+        rdvc_widgets['subvol_points_spinbox'].setValue(10000)
         rdvc_widgets['subvol_points_spinbox'].setToolTip(subvol_points_text)
 
         singleRun_groupBoxFormLayout.setWidget(widgetno, QFormLayout.FieldRole, rdvc_widgets['subvol_points_spinbox'])
@@ -3913,7 +4038,6 @@ This parameter has a strong effect on computation time, so be careful."
         self.rdvc_widgets = rdvc_widgets
 
     def _set_num_points_in_run_to_all(self):
-        int(self.pc_no_points)
         if hasattr(self, 'pc_no_points'):
             maxpoints = int(self.pc_no_points)
         else:
@@ -3961,16 +4085,20 @@ This parameter has a strong effect on computation time, so be careful."
             self.warningDialog("Complete image registration first.", "Error")
             return
 
-        if self.singleRun_groupBox.isVisible():
+        if self.pointcloud_is == 'loaded':
             if not self.roi:
                 self.warningDialog(window_title="Error", 
                                message="Create or load a pointcloud on the viewer first." )
                 return
-        else:
+        elif self.pointcloud_is == 'generated':
             if not self.mask_reader:
                 self.warningDialog(window_title="Error", 
                                message="Load a mask on the viewer first" )
                 return
+        else:
+            self.warningDialog(window_title="Error", 
+                               message="Missing pointcloud_is parameter! Contact developers" )
+            return
         
         run_name = str( self.rdvc_widgets['name_entry'].text() )
         saved_run_names = []
@@ -4066,9 +4194,17 @@ This parameter has a strong effect on computation time, so be careful."
                     #print(subvol_size)
                     subvol_size_count+=1
                     filename = os.path.join( run_folder , "_{}.roi".format(str(subvol_size)))
-                    #print(filename)
-                    if not self.createPointCloud(filename=filename, subvol_size=int(subvol_size)):
-                        return ("pointcloud error")
+                    if self.pointcloud_is == 'generated':
+                        # we will generate the same pointcloud for each subvolume size
+                        pc_subvol_size = self.pointcloud_parameters['pointcloud_size_entry'].text()
+                        if not self.createPointCloud(filename=filename, subvol_size=int(pc_subvol_size)):
+                            return ("pointcloud error")
+                    elif self.pointcloud_is == 'loaded':
+                        # we will use the file with the pointcloud for each subvol size
+                        shutil.copyfile(self.roi, filename)
+                    else:
+                        raise ValueError("pointcloud_is must be either 'generated' or 'loaded'")
+
                     self.roi_files.append(filename)
                     progress_callback.emit(subvol_size_count/len(self.subvol_sizes)*90)
                 #print("finished making pointclouds")
@@ -4263,7 +4399,7 @@ The dimensionality of the pointcloud can also be changed in the Point Cloud pane
         formLayout.setWidget(widgetno, QFormLayout.LabelRole, result_widgets['vec_label'])
 
         result_widgets['vec_entry'] = QComboBox(groupBox)
-        result_widgets['vec_entry'].addItems(['None', 'Total Displacement', 'Displacement with respect to Reference Point 0'])
+        result_widgets['vec_entry'].addItems(['Pointcloud', 'Total Displacement', 'Displacement with respect to Reference Point 0'])
         result_widgets['vec_entry'].currentIndexChanged.connect(self._DVCResultsDisableRanges)
         formLayout.setWidget(widgetno, QFormLayout.FieldRole, result_widgets['vec_entry'])
         widgetno += 1
@@ -4402,12 +4538,15 @@ The dimensionality of the pointcloud can also be changed in the Point Cloud pane
             #print("New roi is", self.roi)
             self.results_folder = results_folder
 
-            if (self.result_widgets['vec_entry'].currentText() == "None"):
+            if (self.result_widgets['vec_entry'].currentText() == "Pointcloud"):
                 self.PointCloudWorker("load pointcloud file")
                 self._removeColormap()
                 # reset the interface
                 self.result_widgets['range_vectors_max_entry'].setEnabled(False)
                 self.result_widgets['range_vectors_min_entry'].setEnabled(False)
+                # set the label of the number of points to the number of points in the pointcloud
+                self.num_processed_points = int( self.pc_no_points )
+                self.DisplayNumberOfPointcloudPoints()
 
             else: 
                 # print("Result list", self.result_list, len(self.result_list))
@@ -4882,7 +5021,14 @@ The dimensionality of the pointcloud can also be changed in the Point Cloud pane
             dialog.Ok.clicked.connect(self.load_session_load)
             dialog.Cancel.clicked.connect(self.load_session_new)
             self.SessionSelectionWindow = dialog
-            dialog.exec()
+            # Try to centre the load session window
+            # from PySide2.QtGui import QScreen
+            # geom = QScreen().availableGeometry()
+            # print (geom)
+            # centrex = geom.topRight().x() + geom.topLeft().x()
+            # centrey = geom.topRight().y() + geom.bottomRight().y()
+            # dialog.move(centrex/2,centrey/2)
+            dialog.open()
         
 
     def load_session_load(self):
@@ -5146,7 +5292,10 @@ Please select the new location of the file, or move it back to where it was orig
 
         else:
             self.subvolume_points = None 
-            self.rdvc_widgets['subvol_points_spinbox'].setValue(self.rdvc_widgets['subvol_points_spinbox'].minimum())
+            # set default to 10000 instead of minimum
+            # value = self.rdvc_widgets['subvol_points_spinbox'].minimum()
+            value = 10000
+            self.rdvc_widgets['subvol_points_spinbox'].setValue(value)
 
             self.points = None
             self.rdvc_widgets['run_points_spinbox'].setValue(self.rdvc_widgets['run_points_spinbox'].minimum())
@@ -5235,10 +5384,11 @@ Please select the new location of the file, or move it back to where it was orig
 
 # Loading and Error windows:
     def progress(self, value):
-        # print("progress emitted")
-        if int(value) > self.progress_window.value():
-            self.progress_window.setValue(value)
-
+        # print("progress emitted", value)
+        if int(value) == 0:
+            value = 1
+        self.progress_window.setValue(value)
+        
 
 
 class SettingsWindow(QDialog):